Unify the var_type enum
Previously, the var_type enum was sensitive to any changes in ordering,
which would break the database.

Additionally, extra preprocessor definitions are used interchangably
with some of the var_type enum, as if they were values for var_type.
This generates literally hundreds of warnings with -Wextra, along with
being less technically correct, and making static analysis regarding
var_type more difficult.

To fix this, explicitly give each element in var_type a value, which
should match the previous value, and move the preprocessor definitions
into the enum as well.