Restrict matching command verbs on player objects
Command verbs are verbs which are not "this none this" and are intended
to be called when a player types a command.

It is a common practice for programmers to write utility verbs on
themselves, often with a dobj/iobj specification of "any" by necessity.

However, another player can issue '@some-test-verb otherplayer', which
will match a otherplayer:@some-test-verb with "any none none".

Generally this is undesirable behavior, and when a verb on another
player is explicitly meant to be called like this, it has "this" as the
matching argument in its verb spec.

Instead, when looking up a verb for a command, if a candidate object is
another player, ensure that that player was mached by a "this" argument,
rather than only "any" arguments.

This should only affect verbs on players intended to be called as
commands by other players without a "this" argument for the target
player, and I'm not really aware of any such verbs or a use for such a
verb that wouldn't use "this" as its specifier anyway.