Add a builtin for calling a verb by verb number
When a verb match is ambiguous, the engine allows the database code to
handle verb lookup.

However, in the case where there may be multiple verbs of the same name,
there is no way to disambiguate them.

For example, consider a book with the following verbs:

- book:turn this to any
- book:turn this

If you have a single book, the engine can find both verbs just fine, but
if you have two books, the database code will have to find the proper
match.

However, there previously was no way to call the second verb, as turn
will match the first, and there was no way to call it via verb number.

This adds a new builtin to do so, with similar semantics to call_verb:

  call_verb_index(OBJ verb_location,
                  INT verb_number,
                  STR verb_name,
                  LIST verb_arguments,
                  [, OBJ this])
    => ANY

The verb_location and verb_number will be used to look up the verb,
verb_name is used to set the verb name within the call, and if not
supplied, as with call_verb, this defaults to the caller.