Add a builtin to find all matching verb numbers
In order to speed up the default in-database lookup, add a builtin which
returns all of the verb numbers on a given object which match the given
verb name.

Otherwise, the only option is to iterate over an object's verbs, and
perform the same comparison the engine does, just within a verb
in-database instead, which can be much slower for large numbers of verbs
and objects.